+++ Quicksilver/Code-App/QSPDFWorkflow.m
@@ -16,13 +16,13 @@ int main (int argc, const char * argv[]) {
 	if (proxy) {
 		NSPasteboard * pboard = [NSPasteboard pasteboardWithUniqueName];
 
-		int fileArgs = 1;
+		NSInteger fileArgs = 1;
 		BOOL putOnShelf = NO;
 
 
 
 
-		int i;
+		NSInteger i;
 		NSMutableArray *filenames = [NSMutableArray arrayWithCapacity:argc-1];
 		NSFileManager *manager = [NSFileManager defaultManager];
 		NSString *currentPath = [manager currentDirectoryPath];
