+++ Quicksilver/Code-QuickStepCore/QSActionProvider.h
@@ -1,5 +1,3 @@
-
-
 #import <Foundation/Foundation.h>
 
 #import "QSAction.h"
@@ -12,7 +10,7 @@
 - (NSArray *)fileTypes;
 - (NSArray *)actions;
 - (NSString *)titleForAction:(NSString *)action;
-- (int)argumentCountForAction:(NSString *)action;
+- (NSInteger)argumentCountForAction:(NSString *)action;
 - (NSImage *)iconForAction:(NSString *)action;
 - (NSArray *)validActionsForDirectObject:(QSObject *)dObject indirectObject:(QSObject *)iObject;
 - (NSArray *)validIndirectObjectsForAction:(NSString *)action directObject:(QSObject *)dObject;
